scripts/requirements.txt
requests
tqdm
python-dotenv
pymongo